Refactored methods createPrimaryKey, addPrimaryKey, addForeignKeys

+++ src/Database/QueryBuilder/CommonQueryBuilder.php
@@ -48,15 +48,36 @@ abstract class CommonQueryBuilder
         return $query;
     }
     
+    protected function createPrimaryKey(Table $table)
+    {
+        if (empty($table->getPrimaryColumns())) {
+            return '';
+        }
+        return $this->primaryKeyString($table);
+    }
+    
+    protected function addPrimaryKey(Table $table)
+    {
+        $queries = [];
+        $primaryColumns = $table->getPrimaryColumns();
+        if (!empty($primaryColumns)) {
+            $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ADD ' . $this->primaryKeyString($table) . ';';
+        }
+        return $queries;
+    }
+    
     protected function dropIndexes(Table $table)
     {
+        if (empty($table->getIndexesToDrop())) {
+            return [];
+        }
         $query = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ';
         $indexes = [];
         foreach ($table->getIndexesToDrop() as $index) {
             $indexes[] = 'DROP INDEX ' . $this->escapeString($index);
         }
         $query .= implode(',', $indexes) . ';';
-        return $query;
+        return [$query];
     }
     
     protected function dropColumns(Table $table)
@@ -69,6 +90,7 @@ abstract class CommonQueryBuilder
         $query .= implode(',', $columns) . ';';
         return $query;
     }
+    
     
     protected function createForeignKeys(Table $table)
     {
@@ -105,6 +127,15 @@ abstract class CommonQueryBuilder
         return $constraint;
     }
     
+    protected function addForeignKeys(Table $table)
+    {
+        $queries = [];
+        foreach ($table->getForeignKeys() as $foreignKey) {
+            $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ADD ' . $this->createForeignKey($foreignKey, $table) . ';';
+        }
+        return $queries;
+    }
+
     abstract public function escapeString($string);
     
     protected function escapeArray(array $array)
@@ -116,5 +147,5 @@ abstract class CommonQueryBuilder
     
     abstract protected function createColumn(Column $column, Table $table);
     
-    abstract protected function createPrimaryKey(Table $table);
+    abstract protected function primaryKeyString(Table $table);
 }
