+++ src/Database/QueryBuilder/PgsqlQueryBuilder.php
@@ -90,11 +90,7 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
      */
     public function alterTable(Table $table)
     {
-        $queries = [];
-        if (!empty($table->getIndexesToDrop())) {
-            $queries[] = $this->dropIndexes($table);
-        }
-        
+        $queries = $this->dropIndexes($table);
         foreach ($table->getForeignKeysToDrop() as $foreignKey) {
             $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' DROP CONSTRAINT ' . $this->escapeString($foreignKey) . ';';
         }
@@ -132,15 +128,8 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
                 $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ALTER COLUMN ' . $this->escapeString($newColumn->getName()) . ' TYPE ' . $this->createType($newColumn) . ' USING ' . $newColumn->getName() . '::' . (isset($this->typeCastMap[$newColumn->getType()]) ? $this->typeCastMap[$newColumn->getType()] : $newColumn->getType()) . ';';
             }
         }
-
-        $primaryColumns = $table->getPrimaryColumns();
-        if (!empty($primaryColumns)) {
-            $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ADD ' . $this->primaryKeyString($table, $primaryColumns) . ';';
-        }
-        
-        foreach ($table->getForeignKeys() as $foreignKey) {
-            $queries[] = 'ALTER TABLE ' . $this->escapeString($table->getName()) . ' ADD ' . $this->createForeignKey($foreignKey, $table) . ';';
-        }
+        $queries = array_merge($queries, $this->addPrimaryKey($table));
+        $queries = array_merge($queries, $this->addForeignKeys($table));
         return $queries;
     }
     
@@ -164,20 +153,11 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
         $col .= $column->allowNull() ? '' : ' NOT NULL';
         return $col;
     }
-    
-    protected function createPrimaryKey(Table $table)
+
+    protected function primaryKeyString(Table $table)
     {
-        return $this->primaryKeyString($table, $table->getPrimaryColumns());
-    }
-    
-    private function primaryKeyString(Table $table, array $primaryColumns = [])
-    {
-        if (empty($primaryColumns)) {
-            return '';
-        }
-        
         $primaryKeys = [];
-        foreach ($primaryColumns as $name) {
+        foreach ($table->getPrimaryColumns() as $name) {
             $primaryKeys[] = $this->escapeString($name);
         }
         return 'CONSTRAINT ' . $this->escapeString($table->getName() . '_pkey') . ' PRIMARY KEY (' . implode(',', $primaryKeys) . ')';
@@ -194,13 +174,16 @@ class PgsqlQueryBuilder extends CommonQueryBuilder implements QueryBuilderInterf
     
     protected function dropIndexes(Table $table)
     {
+        if (empty($table->getIndexesToDrop())) {
+            return [];
+        }
         $query = 'DROP INDEX ';
         $indexes = [];
         foreach ($table->getIndexesToDrop() as $index) {
             $indexes[] = $this->escapeString($index);
         }
         $query .= implode(',', $indexes) . ';';
-        return $query;
+        return [$query];
     }
     
     public function escapeString($string)
